20231126修改提交

+++ ruoyi-system/src/main/java/com/ruoyi/ql/domain/export/WarehousingExportVo.java
@@ -107,7 +107,7 @@ public class WarehousingExportVo {
      * 库存数量
      */
     @ExcelProperty(value = "库存数量")
-    private Long inventoryNumber;
+    private BigDecimal  inventoryNumber;
 
     /**
      * 基准价
