+++ ruoyi-system/src/main/java/com/ruoyi/ql/domain/importvo/WarehousingImportVo.java
@@ -225,7 +225,7 @@ public class WarehousingImportVo {
     /**
      * 库存数量
      */
-    private Long inventoryNumber;
+    private BigDecimal  inventoryNumber;
 
     /**
      * 产品名称
